+++ src/tracker.ts
@@ -87,26 +87,29 @@ export class Tracker extends BaseFileRef {
         if (pipelines.length === 0) {
           return;
         }
-        const pipeline = pipelines[0];
-        const jobs = await gitLab.listPipelineJobs(pipeline.id);
-        if (
-          pipeline.status === "success" &&
-          jobs.find((j) => j.name === "deploy" && j.status === "success")
-        ) {
-          const message = `${projectName} #${issueNumber}: Staging -> ${projectConfig.deployedStatus}`;
-          childProcess.execSync(
-            `osascript -e 'display notification "${message}" with title "Accel Shooter"'`
-          );
-          console.log(message);
-          await clickUp.setTaskStatus(projectConfig.deployedStatus);
-          this.closeItem(projectName, issueNumber);
-        }
-        if (pipeline.status === "failed") {
-          const message = `${projectName} #${issueNumber}: Pipeline failed`;
-          childProcess.execSync(
-            `osascript -e 'display notification "${message}" with title "Accel Shooter"'`
-          );
-          console.log(message);
+        for (const pipeline of pipelines) {
+          const jobs = await gitLab.listPipelineJobs(pipeline.id);
+          const job = jobs.find((j) => j.name === "deploy");
+          if (!job) {
+            continue;
+          }
+          if (pipeline.status === "failed") {
+            const message = `${projectName} #${issueNumber}: Pipeline failed`;
+            childProcess.execSync(
+              `osascript -e 'display notification "${message}" with title "Accel Shooter"'`
+            );
+            console.log(message);
+          }
+          if (job.status === "success") {
+            await clickUp.setTaskStatus(projectConfig.deployedStatus);
+            this.closeItem(projectName, issueNumber);
+            const message = `${projectName} #${issueNumber}: Staging -> ${projectConfig.deployedStatus}`;
+            childProcess.execSync(
+              `osascript -e 'display notification "${message}" with title "Accel Shooter"'`
+            );
+            console.log(message);
+          }
+          break;
         }
       }
     }